ееее все работает ну не все конечно но все
