+++ src/main/java/org/flamierawieo/anirandom/controller/Search.java
@@ -1,11 +1,14 @@
 package org.flamierawieo.anirandom.controller;
 
+import com.mitchellbosecke.pebble.error.PebbleException;
 import org.flamierawieo.anirandom.orm.Anime;
 import org.mongodb.morphia.query.Query;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.servlet.http.HttpServletRequest;
+import java.io.IOException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.regex.Pattern;
@@ -16,8 +19,10 @@ import static org.flamierawieo.anirandom.Util.jsonify;
 @RestController
 public class Search extends Base {
 
-    @RequestMapping("/search")
-    public String get(@RequestParam(value = "s") String searchString) {
+    private static String template = "search.html";
+
+    @RequestMapping("/search.json")
+    public String json(@RequestParam(value = "s") String searchString) {
         if(searchString.length() < 3) {
             return "";
         }
@@ -28,6 +33,11 @@ public class Search extends Base {
         return jsonify(new HashMap() {{
             put("results", animes.stream().map(Anime::toMap).collect(Collectors.toList()));
         }});
+    }
+
+    @RequestMapping("/search")
+    public String get(HttpServletRequest request) throws IOException, PebbleException {
+        return render(template, getContext(request));
     }
 
 }
