+++ src/main/resources/templates/base.html
@@ -11,7 +11,7 @@
   </head>
   <body>
     <div class="anirandom-title">
-      <a href="/">Anirandom</a>
+      <a href="/" class="anirandom-title">Anirandom</a>
     </div>
     {% if authorized %}
       <div class="auth">
