+++ src/main/resources/templates/plan_to_watch.html
@@ -29,15 +29,10 @@
         </section>
       {% endfor %}
     {% else %}
-      <div class="no-text">You didn't add anything yet. <br> You're a bad person and you should be ashamed! When you decide to add something to the list, send it to your friends for them to laugh at how much <br> of a pleb you are, just tell them your username.</div>
-      <div class="no-image"></div>
-      <div class="search">
-        <div class="search-icon"></div>
-        <input type="text" placeholder="priv kak dela suchki">
-        <div class="search-icon"></div>
-        <ul>
-        </ul>
-      </div>
+      <div class="plan_to_watch">Plan to watch</div>
+      <div class="no-text">You haven't added anything to the "Plan to watch" list. You're not a true weeb! A true weeb should watch at least one title a day. Just use the search.</div>
+      <div class="no-image" style="background-image: url(http://data.whicdn.com/images/11666052/original.jpg)"></div>
+      <a href="/search" class="go_to_search">Search</a>
     {% endif %}
   {% else %}
     <div class="no-auth-text"></div>
